requirements.txt
openai==1.2.3
aiogram==3.17.0
requests
asyncpg
python-dotenv
pip==25.0.1
quart==0.19.4
hypercorn==0.14.4

